# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Reverb Collaborative Platform API",
    "version": "1.0.0",
    "description": "API documentation for the Reverb music collaboration project."
  },
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/collaboration/files/{id}/rename": {
      "put": {
        "tags": ["Collaboration"],
        "summary": "Rename an uploaded audio file",
        "description": "Allows the owner of a file to rename it. The file extension will be preserved automatically.",
        "operationId": "renameFile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the file to rename"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["newName"],
                "properties": {
                  "newName": {
                    "type": "string",
                    "example": "MyEditedSong",
                    "description": "New filename (without extension)"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "File renamed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "example": "Renamed"
                    },
                    "file": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "664b21327f5e2e3a9c728c13"
                        },
                        "filename": {
                          "type": "string",
                          "example": "MyEditedSong.wav"
                        },
                        "path": {
                          "type": "string",
                          "example": "/uploads/MyEditedSong.wav"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid new name"
          },
          "403": {
            "description": "User is not authorized to rename the file"
          },
          "404": {
            "description": "File not found"
          }
        }
      }
    },
...
  }
}